refactor(spot): move Connect handler into internal/http
Consolidate delivery in internal/http (package http) alongside the REST
handlers and routes.go, matching the repo convention that handlers live
there. The domain stays pure in internal/spot. The handler struct and
constructor are unexported (newSpotService). Addresses review feedback on
PR #214.

Refs #208
